SignumBase refactoring.
